improve build script
- use ProGuard to shrink and optimize ColorFill code (which doesn't actually improve performance in Hotspot VM)
- libs (designgridlayout and flatlaf) are included without any ProGuard shrinking or optimization (which could cause runtime errors)
- use advzip "insane" zip compression to create a smaller output jar-file (libs are precompressed for a speedup of the build script)

+++ buildjar.sh
@@ -1,17 +1,28 @@
 OUTPUT=colorfill.jar
 TMPDIR=_tmp_
+TMPLIBZIP=lib/_tmplib.zip
 
-rm -f $OUTPUT _$OUTPUT
+if [ ! -f $TMPLIBZIP ]; then
+    rm -rf $TMPDIR
+    mkdir $TMPDIR
+    unzip -q lib/designgridlayout-1.11.jar -x "META-INF/*" -d $TMPDIR
+    unzip -q lib/flatlaf-0.42.jar          -x "META-INF/*" -d $TMPDIR
+    cd $TMPDIR
+    advzip -a -4 -i 1000 ../$TMPLIBZIP *
+    cd ..
+fi
 
+rm -f $OUTPUT _$OUTPUT $OUTPUT.zip
 java -jar $HOME/Programme/proguard6.2/lib/proguard.jar @proguard_config -outjars _$OUTPUT
 
 rm -rf $TMPDIR
 mkdir $TMPDIR
 unzip -q _$OUTPUT -d $TMPDIR
-
 cd $TMPDIR
-advzip -a -4 -i 256 ../$OUTPUT *
+#advzip -a -q -3 ../$OUTPUT.zip *
+advzip -a -4 -i 1000 ../$OUTPUT.zip *
 cd ..
-rm -rf $TMPDIR
+zipmerge $OUTPUT $OUTPUT.zip $TMPLIBZIP
+rm -rf $TMPDIR _$OUTPUT $OUTPUT.zip
 
 chmod u+x $OUTPUT
